Character(전투원) 관련 클래스 자료형 수정

+++ src/main/java/com/dace/dmgr/combat/character/Character.java
@@ -28,9 +28,9 @@ public abstract class Character {
     /** 체력 */
     private final int health;
     /** 이동속도 계수 */
-    private final float speedMultiplier;
+    private final double speedMultiplier;
     /** 히트박스 크기 계수 */
-    private final float hitboxMultiplier;
+    private final double hitboxMultiplier;
 
     /**
      * 액션바에 무기 및 스킬 상태를 표시하기 위한 문자열을 반환한다.
@@ -47,6 +47,7 @@ public abstract class Character {
      * @param i          인덱스
      */
     public void onTick(@NonNull CombatUser combatUser, long i) {
+        // 미사용
     }
 
     /**
@@ -75,6 +76,7 @@ public abstract class Character {
      * @see Character#onAttack(CombatUser, Damageable, int, DamageType, boolean)
      */
     public void onDamage(@NonNull CombatUser victim, Attacker attacker, int damage, @NonNull DamageType damageType, boolean isCrit) {
+        // 미사용
     }
 
     /**
@@ -99,6 +101,7 @@ public abstract class Character {
      * @see Character#onGiveHeal(CombatUser, Healable, int)
      */
     public void onTakeHeal(@NonNull CombatUser target, Healer provider, int amount) {
+        // 미사용
     }
 
     /**
@@ -109,6 +112,7 @@ public abstract class Character {
      * @see Character#onDeath(CombatUser, Attacker)
      */
     public void onKill(@NonNull CombatUser attacker, @NonNull Damageable victim) {
+        // 미사용
     }
 
     /**
@@ -119,6 +123,7 @@ public abstract class Character {
      * @see Character#onKill(CombatUser, Damageable)
      */
     public void onDeath(@NonNull CombatUser victim, Attacker attacker) {
+        // 미사용
     }
 
     /**
